# web/main-ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "inventoryzing",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/scanner/sessions": {
      "post": {
        "summary": "Create Session",
        "operationId": "create_session_api_scanner_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScanSession"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanSessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scanner/sessions/{session_id}": {
      "get": {
        "summary": "Read Session",
        "operationId": "read_session_api_scanner_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanSessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Session",
        "operationId": "delete_session_api_scanner_sessions__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/scans": {
      "post": {
        "summary": "Receive Scan",
        "operationId": "receive_scan_api_agent_scans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanObservation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scanner/sessions/{session_id}/input": {
      "post": {
        "summary": "Manual Scan",
        "operationId": "manual_scan_api_scanner_sessions__session_id__input_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualScanInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentScanResponse": {
        "properties": {
          "outcome": {
            "type": "string",
            "enum": [
              "success",
              "failure",
              "no_action"
            ],
            "title": "Outcome"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "command_completed": {
            "type": "boolean",
            "title": "Command Completed",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "outcome",
          "message"
        ],
        "title": "AgentScanResponse"
      },
...
      "CreateScanSession": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "lookup",
              "move"
            ],
            "title": "Mode"
          },
          "destination_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Id"
          },
          "controller_id": {
            "type": "string",
            "format": "uuid",
            "title": "Controller Id"
          },
          "take_over": {
            "type": "boolean",
            "title": "Take Over",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "controller_id"
        ],
        "title": "CreateScanSession"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualScanInput": {
        "properties": {
          "payload": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Payload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "payload"
        ],
        "title": "ManualScanInput"
      },
...
      "ScanObservation": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "terminal_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal Id"
          },
          "payload": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Payload"
          },
          "source_id": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Source Id"
          },
          "runtime_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Runtime Id"
          },
          "source_type": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Source Type"
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "serial_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Serial Number"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          },
          "symbology": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbology"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "payload",
          "source_id",
          "runtime_id",
          "source_type",
          "received_at"
        ],
        "title": "ScanObservation"
      },
      "ScanResult": {
        "properties": {
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "success",
              "failure",
              "no_action"
            ],
            "title": "Outcome"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "object_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Id"
          },
          "object_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Name"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          }
        },
        "type": "object",
        "required": [
          "sequence",
          "event_id",
          "outcome",
          "message",
          "occurred_at"
        ],
        "title": "ScanResult"
      },
      "ScanSessionView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "terminal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Terminal Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "lookup",
              "move"
            ],
            "title": "Mode"
          },
          "destination_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Id"
          },
          "destination_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Name"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ScanResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "id",
          "terminal_id",
          "mode",
          "destination_id",
          "destination_name",
          "expires_at"
        ],
        "title": "ScanSessionView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}